upstream

+++ SystemMonitor/Network/NetworkInfoController.m
@@ -743,7 +743,7 @@ static void reachabilityCallback(SCNetworkReachabilityRef target, SCNetworkReach
     }
     
 #define ROUNDUP64(a)    \
-    ((a) > 0 ? (1 + (((a) - 1) | (sizeof(uint64_t) - 1))) : sizeof(uint64_t))
+    ((a) > 0 ? (1 + (((a) - 1) | (sizeof(UInt64) - 1))) : sizeof(UInt64))
     
     oxig = xig = (struct xinpgen *)buf;
     for (next = buf + ROUNDUP64(xig->xig_len); next < buf + len; next += ROUNDUP64(xgn->xgn_len))
@@ -778,13 +778,13 @@ static void reachabilityCallback(SCNetworkReachabilityRef target, SCNetworkReach
                     tp = (struct xtcpcb_n *)xgn;
                     break;
                 default:
-                    AMLogWarn(@"unknown kind %d", xgn->xgn_kind);
+                    AMLogWarn(@"unknown kind %ld", (long)xgn->xgn_kind);
                     break;
             }
         }
         else
         {
-            AMLogWarn(@"got %d twice.", xgn->xgn_kind);
+            AMLogWarn(@"got %ld twice.", (long)xgn->xgn_kind);
         }
         
         if ((connectionType == CONNECTION_TYPE_TCP4 && which != ALL_XGN_KIND_TCP) ||
